refactor: move encodeBatchSwap up to setup so it can be shared
