+++ pkg/standalone-utils/test/VaultActions.test.ts
@@ -25,6 +25,7 @@ import {
   encodeJoinPool,
   encodeExitPool,
   encodeSwap,
+  encodeBatchSwap,
   getJoinExitAmounts,
   approveVaultForRelayer,
   PoolKind,
@@ -87,50 +88,6 @@ describe('VaultActions', function () {
 
     poolIdC = await poolC.getPoolId();
   });
-
-  function encodeBatchSwap(params: {
-    swaps: Array<{
-      poolId: string;
-      tokenIn: Token;
-      tokenOut: Token;
-      amount: BigNumberish;
-    }>;
-    outputReferences?: Dictionary<BigNumberish>;
-    sender: Account;
-    recipient?: Account;
-    useInternalBalance?: boolean;
-  }): string {
-    const outputReferences = Object.entries(params.outputReferences ?? {}).map(([symbol, key]) => ({
-      index: tokens.findIndexBySymbol(symbol),
-      key,
-    }));
-
-    if (params.useInternalBalance == undefined) {
-      params.useInternalBalance = false;
-    }
-
-    return relayerLibrary.interface.encodeFunctionData('batchSwap', [
-      SwapKind.GivenIn,
-      params.swaps.map((swap) => ({
-        poolId: swap.poolId,
-        assetInIndex: tokens.indexOf(swap.tokenIn),
-        assetOutIndex: tokens.indexOf(swap.tokenOut),
-        amount: swap.amount,
-        userData: '0x',
-      })),
-      tokens.addresses,
-      {
-        sender: TypesConverter.toAddress(params.sender),
-        recipient: params.recipient ?? TypesConverter.toAddress(recipient),
-        fromInternalBalance: params.useInternalBalance,
-        toInternalBalance: params.useInternalBalance,
-      },
-      new Array(tokens.length).fill(MAX_INT256),
-      MAX_UINT256,
-      0,
-      outputReferences,
-    ]);
-  }
 
   function encodeManageUserBalance(params: {
     ops: Array<{
@@ -332,7 +289,7 @@ describe('VaultActions', function () {
     context('when caller is not authorized', () => {
       it('reverts', async () => {
         await expect(
-          relayer.connect(other).multicall([encodeBatchSwap({ swaps: [], sender: user.address })])
+          relayer.connect(other).multicall([encodeBatchSwap({ relayerLibrary, tokens, swaps: [], sender: user.address, recipient })])
         ).to.be.revertedWith('Incorrect sender');
       });
     });
@@ -365,11 +322,14 @@ describe('VaultActions', function () {
             () =>
               relayer.connect(user).multicall([
                 encodeBatchSwap({
+                  relayerLibrary,
+                  tokens,
                   swaps: [
                     { poolId: poolIdA, tokenIn: tokens.DAI, tokenOut: tokens.MKR, amount: amountInA },
                     { poolId: poolIdC, tokenIn: tokens.SNX, tokenOut: tokens.BAT, amount: amountInC },
                   ],
                   sender,
+                  recipient,
                 }),
               ]),
             tokens,
@@ -406,11 +366,14 @@ describe('VaultActions', function () {
           const receipt = await (
             await relayer.connect(user).multicall([
               encodeBatchSwap({
+                relayerLibrary,
+                tokens,
                 swaps: [
                   { poolId: poolIdA, tokenIn: tokens.DAI, tokenOut: tokens.MKR, amount: amountInA },
                   { poolId: poolIdC, tokenIn: tokens.SNX, tokenOut: tokens.BAT, amount: amountInC },
                 ],
                 sender,
+                recipient,
                 outputReferences: {
                   MKR: toChainedReference(0),
                   SNX: toChainedReference(1),
@@ -440,6 +403,8 @@ describe('VaultActions', function () {
           const receipt = await (
             await relayer.connect(user).multicall([
               encodeBatchSwap({
+                relayerLibrary,
+                tokens,
                 swaps: [
                   { poolId: poolIdA, tokenIn: tokens.DAI, tokenOut: tokens.MKR, amount: amountInA },
                   {
@@ -450,6 +415,7 @@ describe('VaultActions', function () {
                   },
                 ],
                 sender,
+                recipient,
               }),
             ])
           ).wait();
@@ -466,6 +432,8 @@ describe('VaultActions', function () {
               () =>
                 relayer.connect(user).multicall([
                   encodeBatchSwap({
+                    relayerLibrary,
+                    tokens,
                     swaps: [
                       { poolId: poolIdA, tokenIn: tokens.DAI, tokenOut: tokens.MKR, amount: amountInA },
                       { poolId: poolIdC, tokenIn: tokens.SNX, tokenOut: tokens.BAT, amount: amountInC },
@@ -478,6 +446,8 @@ describe('VaultActions', function () {
                     recipient: TypesConverter.toAddress(sender), // Override default recipient to chain the output with the next swap.
                   }),
                   encodeBatchSwap({
+                    relayerLibrary,
+                    tokens,
                     swaps: [
                       // Swap previously acquired MKR for SNX
                       {
@@ -495,6 +465,7 @@ describe('VaultActions', function () {
                       },
                     ],
                     sender,
+                    recipient,
                   }),
                 ]),
               tokens,
@@ -1473,6 +1444,8 @@ describe('VaultActions', function () {
                 recipient: relayer.address,
               }),
               encodeBatchSwap({
+                relayerLibrary,
+                tokens,
                 swaps: [{ poolId: poolIdB, tokenIn: tokens.MKR, tokenOut: tokens.SNX, amount: toChainedReference(1) }],
                 outputReferences: {
                   SNX: toChainedReference(1),
